Move cross-repo plugin read description to core
